Safari-611.1.10/LayoutTests/js/dom/modules/module-load-same-module-from-different-entry-point-in-src-expected.txt

2 different module tag will attempt to load the same module with "src" attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS window.moduleExecutedCount is 0
Executing module-load-same-module-from-different-entry-point.js.
PASS window.moduleExecutedCount is 1
PASS window.moduleExecutedCount is 1
PASS counter is 2
PASS successfullyParsed is true

TEST COMPLETE
